pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-financial-brain"
version = "1.0.0"
description = "AI Financial Brain — Autonomous Analyst + Personal CFO"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",
    "langgraph>=0.2.40",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "weaviate-client>=4.9.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.0",
    "asyncpg>=0.29.0",
    "sentence-transformers>=3.1.0",
    "rank-bm25>=0.2.2",
    "cohere>=5.9.0",
    "torch>=2.4.0",
    "transformers>=4.44.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "arch>=7.1.0",
    "PyPortfolioOpt>=1.5.0",
    "shap>=0.46.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",
    "statsmodels>=0.14.0",
    "mlflow>=2.16.0",
    "prefect>=3.0.0",
    "evidently>=0.4.30",
    "prometheus-client>=0.20.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "redis>=5.1.0",
    "mem0ai>=0.1.0",
    "yfinance>=0.2.40",
    "nsepy>=0.6.1",
    "newsapi-python>=0.2.7",
    "alpha-vantage>=3.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "celery>=5.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[project.scripts]
ai-brain = "main:main"
